Move UDP discovery classes to new udp package
