+++ proxy-robots-module/src/test/java/orwell/proxy/udp/UdpBeaconFinderFactoryTest.java
@@ -1,4 +1,4 @@
-package orwell.proxy;
+package orwell.proxy.udp;
 
 import org.easymock.TestSubject;
 import org.junit.After;
